Restored _application_externalProps to expose all properties
Waveform property lists will now include the assembly controller
properties as well as anything listed in the external properties.
Previous commit's removal of this mechanism was actually because
it was bugged to appear to be the same as _getPropertySet since
the method wasn't modifying the emitted property IDs to the
external IDs, which were then filtered out when compared with
app.query([]).