Changed default to non-decimating
By having this set to non-zero, the SRI is always indicated as 'changed'
which can wreak havoc at downstream users.  An internal issue has been
created to retain the modified SRI when the limiter is in use so that
'changed' is relative to the previously modified SRI, not the incoming.

Related to ESRA-120.