Removed redundant test
The various asserts, etc. in the `_launch` and `_release` methods
effectively check if the application did indeed launch, was seen
in the applications list, and then upon release that the specific
app ID is no longer in the applications list.  So by virtue of the
other tests calling those methods repeatedly, this big test can
be eliminated.