Changed to manual 'deploy' of docker image
DockerHub does not support deploy keys or anything quite like it,
only short-lived tokens returned once you log in with a user/pass
combination, which means the CI setup would have to expose those
keys for an actual user on their site that has write access to the
image in question.  The work-around right now is to make the deploy
create the tagged image and push it to the local repo.  "Someone"
will need to set their notifications to hear successful pipeline
completions and do the work of then pushing it to the public repo
and cleaning the internal registry.

Also changed the 'clean' script since the variables aren't expanding
quite as expected.